Create inference object
